requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
tortoise-orm==0.21.5